HW08 in progress

+++ src/main/webapp/resources/js/topjava.common.js
@@ -4,16 +4,9 @@ function makeEditable(datatableApi) {
     ctx.datatableApi = datatableApi;
     form = $('#detailsForm');
 
-
-    $('#datatable').on('click', 'delete', function () {
-        if (confirm('Are you sure?')) {
-            deleteRow($(this).closest('tr').attr("id"));
-        }
-    });
-
-    $(document).ajaxError(function (event, jqXHR, options, jsExc) {
-        failNoty(jqXHR);
-    });
+    $('.delete').click(function () {
+        deleteRow($(this).attr("id"));
+    })
 
     // solve problem with cache in IE: https://stackoverflow.com/a/4303862/548473
     $.ajaxSetup({cache: false});
@@ -25,34 +18,40 @@ function add() {
 }
 
 function deleteRow(id) {
-    $.ajax({
-        url: ctx.ajaxUrl + id,
-        type: "DELETE"
-    }).done(function () {
-        updateTableWithGet();
-        successNoty("Deleted");
-    });
+    if (confirm('Are you sure?')) {
+        $.ajax({
+            url: ctx.ajaxUrl + id,
+            type: "DELETE"
+        }).done(function () {
+            successNoty("Deleted");
+        });
+    }
 }
 
-function updateTableWithGet() {
+function editRow(id) {
+}
+
+
+function updateTableByGet() {
     $.get(ctx.ajaxUrl, function (data) {
         ctx.datatableApi.clear().rows.add(data).draw();
     });
 }
 
-function updateTableData(data) {
+function updateTableByData(data) {
     ctx.datatableApi.clear().rows.add(data).draw();
+    // ctx.datatableApi.clear().rows.add(data).draw();
 }
 
 function save() {
     $.ajax({
         type: "POST",
         url: ctx.ajaxUrl,
-        contentType: 'application/json',
+        // contentType: 'application/json',
         data: form.serialize()
     }).done(function () {
         $("#editRow").modal("hide");
-        updateTableWithGet();
+        updateTableByGet();
         successNoty("Saved");
     });
 }
